update:支持组件化路由
