--- annotation/src/main/java/com/bravestsnail/annotation/Router.kt
@@ -1,5 +1,0 @@
-package com.bravestsnail.annotation
-
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.SOURCE)
-annotation class Router(val url: String)
