+++ app/src/main/java/com/bravestsnail/router/BaseApplication.kt
@@ -1,0 +1,11 @@
+package com.bravestsnail.router
+
+import android.app.Application
+import com.bravestsnail.router.handler.RouterUtils
+
+class BaseApplication: Application() {
+    override fun onCreate() {
+        super.onCreate()
+        RouterUtils.init(this)
+    }
+}
