+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,4 @@
-#Mon Oct 24 12:27:14 CST 2022
+#Thu Oct 27 00:34:01 CST 2022
 distributionBase=GRADLE_USER_HOME
 distributionUrl=https\://services.gradle.org/distributions/gradle-7.4-bin.zip
 distributionPath=wrapper/dists
