+++ processor/build.gradle
@@ -9,7 +9,6 @@ java {
 }
 
 dependencies {
-    implementation project(':annotation')
     //生成Kotlin代码
     implementation "com.squareup:kotlinpoet:1.6.0"
 }
