+++ processor/src/main/java/com/bravestsnail/router/annotation/Router.kt
@@ -1,0 +1,5 @@
+package com.bravestsnail.router.annotation
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.SOURCE)
+annotation class Router(val url: String)
