+++ processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-com.bravestsnail.processor.RouterProcessor
+com.bravestsnail.router.processor.RouterProcessor
