+++ routerhandler/build.gradle
@@ -38,7 +38,4 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
-
-    api project(':annotation')
-    implementation project(':processor')
 }
