+++ routerhandler/src/main/java/com/bravestsnail/router/handler/RouterAnnotationHandler.kt
@@ -1,0 +1,11 @@
+package com.bravestsnail.router.handler
+
+interface RouterAnnotationHandler {
+    object Default : RouterAnnotationHandler {
+        override fun register(url: String, target: String) {
+            RouterUtils.register(url, target)
+        }
+    }
+
+    fun register(url: String, target: String)
+}
