+++ routerhandler/src/main/java/com/bravestsnail/router/handler/RouterAnnotationInit.kt
@@ -1,0 +1,6 @@
+package com.bravestsnail.router.handler
+
+interface RouterAnnotationInit {
+
+    fun init(handler: RouterAnnotationHandler)
+}
