--- routerhandler/src/main/java/com/bravestsnail/routerhandler/RouterAnnotationInit.kt
@@ -1,6 +1,0 @@
-package com.bravestsnail.routerhandler
-
-interface RouterAnnotationInit {
-
-    fun init(handler: RouterAnnotationHandler)
-}
