+++ test/build.gradle
@@ -1,0 +1,91 @@
+plugins {
+    id 'com.android.library'
+    id 'org.jetbrains.kotlin.android'
+    id 'kotlin-kapt'
+    id 'maven-publish'
+}
+
+android {
+    namespace 'com.bravestsnail.test'
+    compileSdk 32
+
+    defaultConfig {
+//        applicationId "com.bravestsnail.test"
+        minSdk 23
+        targetSdk 32
+        versionCode 1
+        versionName "1.0"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
+    publishing {
+        singleVariant('release') {
+            withSourcesJar()
+        }
+    }
+}
+
+dependencies {
+
+    implementation 'androidx.core:core-ktx:1.7.0'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'com.google.android.material:material:1.5.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    implementation project(':routerhandler')
+    implementation project(':processor')
+    kapt project(':processor')
+}
+
+//task generateSourcesJar(type: Jar) {
+//    from android.sourceSets.main.java.getSrcDirs() // 源码路径
+//    archiveClassifier= "sources"
+//}
+
+def versionName = "1.0.1"
+
+afterEvaluate{
+    publishing{
+        repositories {
+            maven{
+                //自定义本地仓库
+                url="C://tmp"
+            }
+//        // 定义一个 maven 仓库
+//        maven {
+//            //在这里放远程仓库地址
+//            url = "https://xxxxxxxx"
+//            // 仓库用户名密码
+//            credentials {
+//                username = "root"
+//                password = "root"
+//            }
+//        }
+        }
+        publications{
+            maven(MavenPublication) {
+                //这个 components 表示 library构建成功后对应的Android组件
+                from components.release // 上传源码
+                groupId =  "com.bravestsnail.router"
+                artifactId = "librarya"
+                version = versionName
+            }
+        }
+    }
+}
