+++ test/src/main/java/com/bravestsnail/test/MainActivity2.kt
@@ -1,13 +1,13 @@
-package com.bravestsnail.router
+package com.bravestsnail.test
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.bravestsnail.router.annotation.Router
 
-@Router("/test")
-class TestActivity : AppCompatActivity() {
+@Router("/test2")
+class MainActivity2 : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_test)
+        setContentView(R.layout.activity_main2)
     }
 }
